dsp-fastapi/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
itsdangerous
openai
python-dotenv
starlette
lxml
pathlib
pyyaml